added RecyclerView and RecyclerAdapter
Note: AvailabilityRecyclerView.java is an attempt to use RecyclerView instead of ListView, because the former is more modern and more easier to implement. That being said, I haven't deleted the old files using ListView since I'm not sure if the RecyclerView is working. The main activity was modified to test AvailabilityRecyclerView, but due to config failures onClick wasn't working. Since it's 2:53am rn im gonna commit before sleeping.