Clean up login activity a little more
Since this class in particular will be scrutinized by a marker

I was initially trying to fix a bug where I couldn't sign in with most of the users in the database, seemingly randomly. (Seemed like some race condition.) I'm unfortunately not sure if that bug still lingers; login seems to work for the moment anyway.